refactor(api): move secret logic to separate package
